[Patchset 5] trace_processor: make args table a lot more efficient
Before, we were doing a full table scan every time we joined the
counters (or any future table) with the args table. Instead, since we
have the multimap of id -> row index, we can do an O(1) lookup to find
the rows associated with the id (if they exist).

This makes a full inner join a lot faster:
Old code on a 6MB trace: 436s
New code on a 6MB trace: 300ms

+++ src/trace_processor/args_table.cc
@@ -91,7 +91,7 @@ void ArgsTable::IdColumn::Filter(int op,
   auto id = sqlite_utils::ExtractSqliteValue<RowId>(value);
   const auto& args_for_id = storage_->args().args_for_id();
   auto it_pair = args_for_id.equal_range(id);
-  std::vector<size_t> rows;
+  std::vector<uint32_t> rows;
   for (auto it = it_pair.first; it != it_pair.second; it++) {
     rows.push_back(it->second);
   }
